// frontend/src/api/models/handlerPayablePaymentRecordResponse.ts
/**
 * // This file is auto-generated by orval
 * // Do not edit manually
 *
 */

/**
 * Payable payment record response
 */
export interface HandlerPayablePaymentRecordResponse {
  amount?: number
  /** @Description Timestamp when the payment was applied in ISO 8601 format
@Example 2026-01-25T14:00:00Z */
  applied_at?: string
  id?: string
  payment_voucher_id?: string
  remark?: string
}